backend/requirements.txt
Django==4.2.4
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
gunicorn==21.2.0